+++ src/entities/notesubfolder.cpp
@@ -1,6 +1,7 @@
 #include "entities/notesubfolder.h"
 #include "notefolder.h"
 #include "note.h"
+#include "tag.h"
 #include <QDebug>
 #include <QSqlRecord>
 #include <QSqlError>
@@ -192,8 +193,15 @@ bool NoteSubFolder::rename(QString newName) {
 
     if (dir.exists() && !newName.isEmpty()) {
         QString oldPath = fullPath();
+        QString oldRelativePath = relativePath();
         setName(newName);
         QString newPath = fullPath();
+        QString newRelativePath = relativePath();
+
+        // rename the note sub folder paths of note tag links
+        // (needs to be done before the folder rename because folder renaming
+        // will cause a reload which would trigger the removal of the tag links)
+        Tag::renameNoteSubFolderPathsOfLinks(oldRelativePath, newRelativePath);
 
         // rename the note subfolder
         return dir.rename(oldPath, newPath);