+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 19.3.1
 - added some markdown text edit fixes and changes
   (for [QMarkdownTextEdit #21](https://github.com/pbek/qmarkdowntextedit/pull/21), thank you @sandsmark)
+- implemented preserving of note tag links when renaming note sub-folders
+  (for [#1166](https://github.com/pbek/QOwnNotes/issues/1166))
 
 ## 19.3.0
 - added **Galician translation** and **Norwegian translation** (a big thank you